# api/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/settings/smtp/test": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需管理员；使用当前数据库中的 SMTP 配置向指定邮箱发送测试邮件，不写库",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "测试 SMTP 邮件发送",
                "parameters": [
                    {
                        "description": "收件人邮箱",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.TestSMTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/settings/{group}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需管理员；按 group 读取设置键值对（SMTP 密码脱敏）；group 可为 general/seo/smtp/comment/security",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "读取站点设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设置分组",
                        "name": "group",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SettingsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需管理员；批量 upsert 指定 group 的设置；SMTP 密码传 \"***\" 表示不修改",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "更新站点设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设置分组",
                        "name": "group",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "设置键值对",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.UpdateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SettingsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需管理员；返回文章/用户/评论总数、今日访问量及热门文章 Top 10",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "站点统计数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SiteStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/articles/{id}/like": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登录用户对文章点赞；同一文章只能点赞一次，重复点赞返回 409",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "文章点赞",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登录用户取消对文章的点赞；未点赞时返回 404",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "取消文章点赞",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.ArticleStatItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "likeCount": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "viewCount": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SettingsResponse": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "settings": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.SiteStatsResponse": {
            "type": "object",
            "properties": {
                "articleCount": {
                    "type": "integer"
                },
                "commentCount": {
                    "type": "integer"
                },
                "todayViews": {
                    "type": "integer"
                },
                "topArticles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.ArticleStatItem"
                    }
                },
                "userCount": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.TestSMTPRequest": {
            "type": "object",
            "required": [
                "to"
            ],
            "properties": {
                "to": {
                    "type": "string"
                }
            }
        },
...
        "github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_api_v1.UpdateSettingsRequest": {
            "type": "object",
            "required": [
                "settings"
            ],
            "properties": {
                "settings": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_HappyLadySauce_Beehive-Blog_cmd_app_types_common.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}